Реализованa storage_set_by_index

+++ libs/macro/storage.h
@@ -101,7 +101,7 @@ int storage_add_arg_by_index(storage *const stg, const size_t id, const size_t i
  *	@param	id			Macro name
  *	@param	value		Macro replacement
  *
- *	@return	Index of record, @c SIZE_MAX on failure
+ *	@return	Index of new record, @c SIZE_MAX on failure
  */
 inline size_t storage_set(storage *const stg, const char32_t *const id, const char32_t *value)
 {
@@ -115,9 +115,9 @@ inline size_t storage_set(storage *const stg, const char32_t *const id, const ch
  *	@param	id			Index of record
  *	@param	value		Macro replacement
  *
- *	@return	@c 0 on success, @c -1 on failure
+ *	@return	Index of new record, @c SIZE_MAX on failure
  */
-int storage_set_by_index(storage *const stg, const size_t id, const char32_t *value);
+size_t storage_set_by_index(storage *const stg, const size_t id, const char32_t *value);
 
 
 /**
